* 'master' of github.com:FoxComm/highlander: (436 commits)
  Create spec.md
  Create use-cases.md
  Create auto-apply-algorithm.md
  Create use-cases.md
  Create spec.md
  Create auto-apply-algorithm.md
  Create tiered-promos.md
  Fix paths’ resolution in scalafmt scripts & have some Unicode
  Create use-cases.md
  Hotfix for legacy mess
  Delete projects.sh
  Update README.md
  Don't repeat onboarding builds twice
  Modify git describe for Ashes
  Fix search-service Makefile
  Debug message
  Build everything if on master, quit if on branch
  Update builder.sh
  Add npm config registry to fix ECONNRESET during install (#2125)
  Use mesos-consul again, repack base images (#2123)
  ...

+++ phoenix-scala/phoenix/app/phoenix/models/shipping/ShippingMethod.scala
@@ -15,7 +15,7 @@ case class ShippingMethod(id: Int = 0,
                           storefrontDisplayName: String,
                           code: String,
                           shippingCarrierId: Option[Int] = None,
-                          price: Int,
+                          price: Long,
                           isActive: Boolean = true,
                           conditions: Option[QueryStatement] = None,
                           restrictions: Option[QueryStatement] = None)
@@ -46,7 +46,7 @@ class ShippingMethods(tag: Tag) extends FoxTable[ShippingMethod](tag, "shipping_
   def storefrontDisplayName = column[String]("storefront_display_name")
   def code                  = column[String]("code")
   def shippingCarrierId     = column[Option[Int]]("shipping_carrier_id")
-  def price                 = column[Int]("price")
+  def price                 = column[Long]("price")
   def isActive              = column[Boolean]("is_active")
   def conditions            = column[Option[QueryStatement]]("conditions")
   def restrictions          = column[Option[QueryStatement]]("restrictions")
